BIT-1624 Fix OutOfMemoryException when uploading large files (#1293)

+++ app/src/main/java/com/x8bit/bitwarden/data/vault/datasource/network/model/SendJsonRequest.kt
@@ -48,7 +48,7 @@ data class SendJsonRequest(
     val deletionDate: ZonedDateTime,
 
     @SerialName("fileLength")
-    val fileLength: Int?,
+    val fileLength: Long?,
 
     @SerialName("file")
     val file: SyncResponseJson.Send.File?,
